improve auto aim, add for drones
drones also use adjusted target but it does not work the same as bullets, might remove?

+++ Graphics.java
@@ -601,8 +601,8 @@ public class Graphics extends Raylib {
      * @param vec The vector to get the length of
      * @return The length of the vector
      */
-    public static double length(Vector2 vec) {
-        return Math.sqrt(vec.x * vec.x + vec.y * vec.y);
+    public static float length(Vector2 vec) {
+        return (float)Math.sqrt(vec.x * vec.x + vec.y * vec.y);
     }
 
     /**
